+++ Dungeon Manager/InitialViewController.h
@@ -16,6 +16,6 @@
 
 @property (weak, nonatomic) IBOutlet UICollectionView *campaignCollection;
 
--(void)deleteCampaign: (int)campaignNumber;
+-(void)deleteCampaign: (NSInteger)campaignNumber;
 
 @end
